Add fetch every 3 seconds features for Individual
